+++ maps/mapComplex.h
@@ -6,13 +6,13 @@
 
 struct MapComplex {
 	char key[KEY_SIZE];
-	double complex value;
+	long double complex value;
 	short empty;
 };
 
 void initMapComplex(struct MapComplex m[MAP_SIZE]);
 
-void insertComplex(struct MapComplex m[MAP_SIZE], char key[KEY_SIZE], double complex value);
+void insertComplex(struct MapComplex m[MAP_SIZE], char key[KEY_SIZE], long double complex value);
 unsigned int findComplex(struct MapComplex *m, char key[KEY_SIZE]);
 
 #endif //CALCULATORPROJECT_MAPCOMPLEX_H
